fix(audio): keep media-session rate conversion out of the audio ISR (#100)
* fix(audio): keep media-session rate conversion out of the audio ISR

Bound media (AX) sessions at non-48 kHz ran the 44.1->48 kHz polyphase
FIR inside audio_fabric_isr. On core 0 that ISR shares the GIC priority
and core with isr_video, so every collision delayed the vblank handler
past the overlay rearm window: the per-vblank overlay VDMA reset landed
in active video, bumping the generation and invalidating both line
buffer banks, which showed as windowed-PIP flicker (zz9000-drivers#83,
reproducible with ZZPlay and any MPEG-1 video whose audio binds to AX;
--audio=ahi and --audio=none were clean, and RC2 with the linear
resampler was clean).

Media sessions now stage converted 48 kHz PCM through the existing
preconvert ring on the main loop, exactly like bound audio streams, so
the ISR fill only copies. The exact view<->session cursor mapping lives
in audio_pump_media_view.h: whole 3840-byte view periods map to whole
source periods (882 frames / 3528 bytes at 44.1 kHz stereo), retirement
is clamped to what the converter really consumed so the zero-padded EOS
tail retires exactly, and the preconvert ring owner is tagged by kind
because media and stream session ids are independent namespaces.

The preconvert source cursors widen to uint64_t so media session
cursors (already 64-bit) map without truncation.

overlay.c gains only a comment: the per-vblank rearm reset is the
overlay stream's phase lock to the raster and must stay unconditional
(a gating experiment confirmed drift artifacts when it is skipped).

Hardware-verified: flicker gone with ZZPlay + AX audio at 1280x1024/32.

* Address PR review feedback (#100)

- Anchor the media view cursors to the session's staged position at
  begin: a resumed session rewinds staging to a nonzero retired cursor,
  and without the anchor the EOS retire clamp mixed an absolute staged
  cursor with a zero-based retired cursor, retiring a full period the
  session must reject and wedging end-of-stream drain.

+++ test/audio/audio_pump_preconvert_test.c
@@ -33,7 +33,7 @@ static void test_44100_matches_converter(void)
 	uint8_t input[882U * 4U];
 	uint8_t ring[RING_BYTES];
 	int16_t expected[960U * 2U];
-	uint32_t consumed = 0U;
+	uint64_t consumed = 0U;
 	uint32_t i;
 
 	for (i = 0U; i < 882U * 2U; i++)
@@ -70,7 +70,7 @@ static void test_48000_is_bit_exact(void)
 	struct audio_pump_preconvert_source source;
 	uint8_t input[AUDIO_PUMP_PRECONVERT_PERIOD_BYTES];
 	uint8_t ring[RING_BYTES];
-	uint32_t consumed = 0U;
+	uint64_t consumed = 0U;
 	uint32_t i;
 
 	for (i = 0U; i < sizeof(input); i++)
@@ -96,7 +96,7 @@ static void test_shortage_waits_and_tail_drains(void)
 	struct audio_pump_preconvert_source source;
 	uint8_t input[100U * 4U];
 	uint8_t ring[RING_BYTES];
-	uint32_t consumed = 0U;
+	uint64_t consumed = 0U;
 
 	memset(input, 0x35, sizeof(input));
 	memset(&source, 0, sizeof(source));
@@ -123,9 +123,9 @@ static void test_cursor_wrap_rebases_both_cursors(void)
 	struct audio_pump_preconvert_source source;
 	uint8_t input[AUDIO_PUMP_PRECONVERT_PERIOD_BYTES];
 	uint8_t ring[RING_BYTES];
-	uint32_t consumed = 0U;
+	uint64_t consumed = 0U;
 	uint32_t base;
-	uint32_t produced_before;
+	uint64_t produced_before;
 	uint32_t staged_before;
 
 	memset(input, 0x5a, sizeof(input));
